// src/types/managerAppearance.ts
export interface ManagerAppearance{
    key_id: number;
    tournament_id: string;
    tournament_name: string;
    match_id: string;
    match_name: string;
    match_date: string;
    group_name: string;
    team_id: string;
    team_name: string;
    home_team: number;
    stage_name: string;
    team_code: string;
    away_team: number;
    manager_id: string;
    family_name: string;
    given_name: string;
    country_name: string;
}